refactor(registrar): extract proof handler
