+++ crates/proof/tee/registrar/src/lib.rs
@@ -38,6 +38,12 @@ pub use metrics::RegistrarMetrics;
 mod prover;
 pub use prover::ProverClient;
 
+mod proof_handler;
+pub use proof_handler::{
+    InFlightRegistrationGuard, ProofHandler, ProofHandlerConfig, ProofHandlerContext,
+    ProofHandlerRequest,
+};
+
 mod registry;
 pub use registry::{RegistryClient, RegistryContractClient};
